refactor: extract scroll helpers to shared utility
Move duplicated scroll-to-center and scroll-to-top logic (3 call sites)
into src/utils/scrollHelpers.ts. No behavior change.
