+++ src/components/CustomPlayButton.tsx
@@ -32,6 +32,7 @@ import {
   logError,
 } from "../api/backend";
 import { getRommConnectionState } from "../utils/connectionState";
+import { scrollToTop } from "../utils/scrollHelpers";
 import { showConflictResolutionModal } from "./ConflictModal";
 import type { DownloadProgressEvent, DownloadCompleteEvent } from "../types";
 
@@ -622,21 +623,8 @@ export const CustomPlayButton: FC<CustomPlayButtonProps> = ({ appId }) => {
           backgroundSize: "330% 100%",
         }}
         onClick={handlePlay}
-        onFocus={(e: any) => {
-          const el = e.currentTarget as HTMLElement;
-          setTimeout(() => {
-            if (!el) return;
-            let scrollParent: HTMLElement | null = el.parentElement;
-            while (scrollParent) {
-              const ov = window.getComputedStyle(scrollParent).overflowY;
-              if (ov === "scroll" || ov === "auto") break;
-              scrollParent = scrollParent.parentElement;
-            }
-            if (scrollParent) {
-              scrollParent.scrollTo({ top: 0, behavior: "smooth" });
-            }
-          }, 50);
-        }}
+        // @ts-expect-error onFocus works at runtime; not in Decky's DialogButton types
+        onFocus={scrollToTop}
       >
         Play
       </DialogButton>
